Consolidated text utils into one file
