+++ src/lib/dragon/builder-states/BuilderHistory.svelte
@@ -8,7 +8,7 @@
 		nextBuilderState,
 		dragonBuilderHistoryMaxLength
 	} from '.';
-	import { capitalizeFirstLetter } from '..';
+	import { capitalizeFirstLetter } from '$lib/text-utils';
 	import type { DragonConfig } from '../dragon-config';
 	import DragonConfigPreview from '../DragonConfigPreview.svelte';
 	import { getToastStore, type ToastSettings } from '@skeletonlabs/skeleton';
