+++ src/lib/dragon/stat-block/actions/breath-weapons/Breath2Blue.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { capitalizeFirstLetter } from '$lib/dragon';
+	import { capitalizeFirstLetter } from '$lib/text-utils';
 	import type { DragonStats } from '$lib/dragon/dragon-stats';
 	import SpellLink from '$lib/spells/SpellLink.svelte';
