+++ src/lib/dragon/stat-block/actions/breath-weapons/Breath2Green.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { capitalizeFirstLetter } from '$lib/dragon';
+	import { capitalizeFirstLetter } from '$lib/text-utils';
 	import type { DragonStats } from '$lib/dragon/dragon-stats';
 
 	export let dragon: DragonStats;
