+++ src/lib/dragon/stat-block/preamble/PreambleAbilityScores.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import type { DragonStats } from '$lib/dragon/dragon-stats';
-	import { ABILITIES, numberWithSign } from '$lib/dragon';
+	import { ABILITIES } from '$lib/dragon';
+	import { numberWithSign } from '$lib/text-utils';
 
 	export let dragon: DragonStats;
 </script>
