+++ src/lib/index.ts
@@ -1,20 +1,4 @@
-import { normalizeString } from './dragon';
-
-// https://stackoverflow.com/a/13627586/17934762
-export function numberWithOrdinalSuffix(i: number): string {
-	const j = Math.abs(i) % 10,
-		k = Math.abs(i) % 100;
-	if (j == 1 && k != 11) {
-		return `${i}st`;
-	}
-	if (j == 2 && k != 12) {
-		return `${i}nd`;
-	}
-	if (j == 3 && k != 13) {
-		return `${i}rd`;
-	}
-	return `${i}th`;
-}
+import { normalizeString } from './text-utils';
 
 export const NAV_LINKS = [
 	{
